fNeuro/pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=44.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fNeuro"
version = "1.0.0"
description = "functions used in neuroimaging analysis"
readme = "README.md"
authors = [{ name = "Daniel Halls", email = "daniel.j.halls@kcl.ac.uk" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["neuroimaging"]

dependencies = [
    'nilearn',
    'scipy',
    'numpy',
    'pandas',
    'python-decouple',
    'scikit-learn',
    'SQLAlchemy'
]
requires-python = ">=3.8"